svg/interact/scripted/defer-01.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>defer on script</title><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#ScriptElement" rel="help"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script>let script_run_status = &quot;inline&quot;; let t = async_test(&quot;defer script runs later&quot;);</script><script defer="defer" href="defer.js" type="text/javascript"/><script>t.step(() =&gt; { assert_equals(script_run_status, &quot;inline&quot;, &quot;script run status&quot;); script_run_status = &quot;deferred&quot;; });</script></svg>
