LayoutTests/imported/w3c/web-platform-tests/svg/interact/scripted/defer-01-expected.txt


FAIL defer script runs later assert_equals: script run status expected "deferred" but got "inline"
